Vuex와 Helper 속성에 따른 모듈화
