+++ src/store/Store.js
@@ -1,5 +1,7 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import Vue from 'vue'
+import Vuex from 'vuex'
+import * as getters from './getters'
+import * as mutations from './mutations'
 
 Vue.use(Vuex);
 
@@ -20,30 +22,6 @@ export const store = new Vuex.Store({
     state: {
         todoItems: storage.fetch() //관리하고 싶은 아이템을 등록한다?
     },
-    getters: {
-      storedTodoItems(state){
-          return state.todoItems;
-      }
-    },
-    mutations: {
-        addOneItem(state, todoItem) {
-            const obj = {completed: false, item: todoItem};
-            localStorage.setItem(todoItem, JSON.stringify(obj))
-            // this.todoItems.push(obj)
-            state.todoItems.push(obj)
-        },
-        removeOneItem(state, payload) {
-            localStorage.removeItem(payload.todoItem.item)
-            state.todoItems.splice(payload.index, 1)
-        },
-        toggleOneItem(state, payload) {
-            state.todoItems[payload.index].completed = !state.todoItems[payload.index].completed
-            localStorage.removeItem(payload.todoItem.item);
-            localStorage.setItem(payload.todoItem.item, JSON.stringify(payload.todoItem))
-        },
-        clearAllItems(state) {
-            localStorage.clear();
-            state.todoItems = [];
-        }
-    }
+    getters: getters,
+    mutations: mutations
 });
