+++ 프로젝트구조화와모듈화.md
@@ -1,0 +1,7 @@
+![image-20200810132514512](https://tva1.sinaimg.cn/large/007S8ZIlgy1ghllohvkf5j30qb0cqwh7.jpg)
+
+Vuex.Store({})  각각의 내용을 모듈화 시킨다면 !?
+
+![image-20200810132608400](https://tva1.sinaimg.cn/large/007S8ZIlgy1ghllpfsh1qj30qb0b2djf.jpg)
+
+각각의 핼퍼 함수를 파일로 나눠서 저장함.
